refactor: exportToString is generic for both rules
