+++ plugin/src/test/java/io/jenkins/plugins/casc/misc/JenkinsConfiguredWithCodeRule.java
@@ -1,16 +1,13 @@
 package io.jenkins.plugins.casc.misc;
 
 import io.jenkins.plugins.casc.ConfigurationAsCode;
-import java.io.ByteArrayOutputStream;
 import java.lang.reflect.Field;
-import java.nio.charset.StandardCharsets;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Objects;
 import java.util.stream.Collectors;
 import org.apache.commons.lang.StringUtils;
 import org.hamcrest.core.StringContains;
-import org.jvnet.hudson.test.JenkinsRule;
 
 import static java.lang.reflect.Modifier.isPublic;
 import static java.lang.reflect.Modifier.isStatic;
@@ -18,7 +15,7 @@ import static java.lang.reflect.Modifier.isStatic;
 /**
  * @author lanwen (Kirill Merkushev)
  */
-public class JenkinsConfiguredWithCodeRule extends JenkinsRule {
+public class JenkinsConfiguredWithCodeRule extends JenkinsConfiguredRule {
 
     @Override
     public void before() throws Throwable {
@@ -77,24 +74,5 @@ public class JenkinsConfiguredWithCodeRule extends JenkinsRule {
             }
         }
         return null;
-    }
-
-    //TODO: Looks like API defect, exception should be thrown
-    /**
-     * Exports the Jenkins configuration to a string.
-     * @return YAML as string
-     * @param strict Fail if any export operation returns error
-     * @throws Exception Export error
-     * @throws AssertionError Failed to export the configuration
-     * @since 1.25
-     */
-    public String exportToString(boolean strict) throws Exception {
-        final ByteArrayOutputStream out = new ByteArrayOutputStream();
-        ConfigurationAsCode.get().export(out);
-        final String s = out.toString(StandardCharsets.UTF_8.name());
-        if (strict && s.contains("Failed to export")) {
-            throw new AssertionError("Failed to export the configuration: " + s);
-        }
-        return s;
     }
 }
