+++ plugin/src/test/java/io/jenkins/plugins/casc/misc/JenkinsConfiguredWithReadmeRule.java
@@ -3,7 +3,6 @@ package io.jenkins.plugins.casc.misc;
 import io.jenkins.plugins.casc.ConfigurationAsCode;
 import org.apache.commons.lang.StringUtils;
 import org.hamcrest.core.StringContains;
-import org.jvnet.hudson.test.JenkinsRule;
 
 import java.io.ByteArrayOutputStream;
 import java.lang.reflect.Field;
@@ -19,7 +18,7 @@ import static java.lang.reflect.Modifier.isStatic;
 /**
  * @author v1v (Victor Martinez)
  */
-public class JenkinsConfiguredWithReadmeRule extends JenkinsRule {
+public class JenkinsConfiguredWithReadmeRule extends JenkinsConfiguredRule {
 
     @Override
     public void before() throws Throwable {
@@ -80,24 +79,5 @@ public class JenkinsConfiguredWithReadmeRule extends JenkinsRule {
             }
         }
         return null;
-    }
-
-    //TODO: Looks like API defect, exception should be thrown
-    /**
-     * Exports the Jenkins configuration to a string.
-     * @return YAML as string
-     * @param strict Fail if any export operation returns error
-     * @throws Exception Export error
-     * @throws AssertionError Failed to export the configuration
-     * @since 1.25
-     */
-    public String exportToString(boolean strict) throws Exception {
-        final ByteArrayOutputStream out = new ByteArrayOutputStream();
-        ConfigurationAsCode.get().export(out);
-        final String s = out.toString(StandardCharsets.UTF_8.name());
-        if (strict && s.contains("Failed to export")) {
-            throw new AssertionError("Failed to export the configuration: " + s);
-        }
-        return s;
     }
 }
